requirements.txt
tqdm==4.62.2
numpy==1.19.5
pydantic==1.10.7
fastapi==0.105.0
Pillow==10.1.0
python-dotenv==1.0.0
ultralytics==8.0.229
uvicorn==0.25.0
